+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/arg_parser.go
@@ -19,11 +19,12 @@ const (
 	serviceIdArgName     = "service_id"
 	serviceConfigArgName = "service_config"
 
-	containerImageNameKey = "container_image_name"
-	usedPortsKey          = "used_ports"
-	entryPointArgsKey     = "entry_point_args"
-	cmdArgsKey            = "cmd_args"
-	envVarArgsKey         = "env_vars"
+	containerImageNameKey         = "container_image_name"
+	usedPortsKey                  = "used_ports"
+	entryPointArgsKey             = "entry_point_args"
+	cmdArgsKey                    = "cmd_args"
+	envVarArgsKey                 = "env_vars"
+	filesArtifactMountDirpathsKey = "files_artifact_mount_dirpaths"
 
 	portNumberKey   = "number"
 	portProtocolKey = "protocol"
@@ -74,6 +75,11 @@ func ParseServiceConfigArg(serviceConfig *starlarkstruct.Struct) (*kurtosis_core
 		return nil, interpretationErr
 	}
 
+	filesArtifactMountDirpaths, interpretationErr := parseFilesArtifactMountDirpaths(serviceConfig)
+	if interpretationErr != nil {
+		return nil, interpretationErr
+	}
+
 	builtConfig := services.NewServiceConfigBuilder(containerImageName).WithPrivatePorts(
 		privatePorts,
 	).WithEntryPointArgs(
@@ -82,6 +88,8 @@ func ParseServiceConfigArg(serviceConfig *starlarkstruct.Struct) (*kurtosis_core
 		cmdArgs,
 	).WithEnvVars(
 		envVars,
+	).WithFilesArtifactMountDirpaths(
+		filesArtifactMountDirpaths,
 	).Build()
 
 	return builtConfig, nil
@@ -238,6 +246,19 @@ func parseEnvVars(serviceConfig *starlarkstruct.Struct) (map[string]string, *sta
 		return nil, interpretationErr
 	}
 	return envVarArgs, nil
+}
+
+func parseFilesArtifactMountDirpaths(serviceConfig *starlarkstruct.Struct) (map[string]string, *startosis_errors.InterpretationError) {
+	_, err := serviceConfig.Attr(filesArtifactMountDirpathsKey)
+	//an error here means that no argument was found which is alright as this is an optional
+	if err != nil {
+		return map[string]string{}, nil
+	}
+	entryPointArgs, interpretationErr := extractMapStringStringValue(serviceConfig, filesArtifactMountDirpathsKey, serviceConfigArgName)
+	if interpretationErr != nil {
+		return nil, interpretationErr
+	}
+	return entryPointArgs, nil
 }
 
 func extractStringValue(structField *starlarkstruct.Struct, key string, argNameForLogging string) (string, *startosis_errors.InterpretationError) {