+++ docs/changelog.md
@@ -3,10 +3,14 @@
 ### Features
 * Added an `exec` command to Startosis
 * Added a `store_files_from_service` command to Startosis
+* Added the ability to pass `files_artifact_mount_dirpaths` to the service config
 * Added a `read_file` command to Startosis
 
 ### Changes
 * Fixed a typo in a filename
+
+### Fixes
+* Fixed a bug in exec where we'd propagate a `nil` error
 
 # 0.51.0
 ### Breaking Changes